// src/lib/types.ts
export type OrderStatus = 'pending' | 'approved' | 'prepared' | 'served' | 'cancelled' | 'billed';
export type UserRole = 'admin' | 'manager' | 'waiter' | 'user';
export type UserStatus = 'pending' | 'approved';
export type BillStatus = 'unpaid' | 'paid';
export type PurchaseStatus = 'ordered' | 'received' | 'cancelled';
export type StockUsageCategory = 'kitchen_prep' | 'spillage' | 'staff_meal' | 'other';
export type OrderType = 'dine-in' | 'pickup';
export type OnlineOrderStatus = 'payment_pending' | 'confirmed' | 'preparing' | 'out_for_delivery' | 'delivered' | 'cancelled';


export interface MenuItemIngredient {
  stockItemId: string;
  quantity: number; // The amount of the stock item used in this menu item
}

export interface MenuItem {
  id: string;
  name: string;
  price: number;
  category: string;
  imageUrl: string;
  isAvailable: boolean;
  ingredients: MenuItemIngredient[];
  costOfGoods: number;
}

export interface OrderItem {
  menuItemId: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: string;
  tableNumber?: number; // Optional for pickup orders
  orderType: OrderType;
  items: OrderItem[];
  status: OrderStatus;
  waiterId: string;
  timestamp: string;
  cancellationReason?: string;
}

export interface Bill {
    id: string;
    tableNumber?: number; // Optional for pickup orders
    orderIds: string[];
    waiterId?: string; // Optional as manager handles billing
    subtotal: number;
    tax: number;
    total: number;
    status: BillStatus;
    timestamp: string;
}

export interface Waiter {
  id: string;
  name: string;
  userId: string;
}

export interface Table {
  id: string;
  tableNumber: number;
  status: 'available' | 'occupied';
  waiterId: string | null;
}